# cmd/api/static/openapi.yaml
openapi: 3.0.3
info:
  title: DriveBai API
  description: |
    DriveBai - Car Marketplace and Rental App API

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Handling
    All errors return a JSON object with an `error` field containing `code` and `message`.

    Error codes:
    - `EMAIL_TAKEN` - Email already registered
    - `INVALID_CREDENTIALS` - Wrong email or password
    - `RATE_LIMITED` - Too many requests
    - `UNAUTHORIZED` - Authentication required
    - `TOKEN_EXPIRED` - Access token expired
    - `TOKEN_INVALID` - Invalid token
    - `INVALID_ROLE` - Invalid role specified
    - `INCOMPLETE_ONBOARDING` - Onboarding requirements not met
  version: 0.2.0
  contact:
    name: DriveBai Support
    email: support@drivebai.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.drivebai.com/api/v1
    description: Production server
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email (deprecated)
      description: |
        **DEPRECATED** - Email verification is no longer required.
        This endpoint returns success for backwards compatibility.
      deprecated: true
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /auth/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access token and refresh token.
        The old refresh token is invalidated (token rotation).
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request password reset
      description: |
        Sends a password reset link to the user's email.
        Always returns 202 to prevent email enumeration.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: |
        Resets the user's password using the token from the reset link.
        All existing sessions are invalidated.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/resend-otp:
    post:
      tags: [Auth]
      summary: Resend OTP code (deprecated)
      description: |
        **DEPRECATED** - OTP is no longer used.
        This endpoint returns success for backwards compatibility.
      deprecated: true
      operationId: resendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOTPRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserProfile'
    VerifyEmailRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: ^\d{6}$
          example: "123456"
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
          description: Token from password reset email
        new_password:
          type: string
          format: password
          minLength: 8
    ResendOTPRequest:
      type: object
      required:
        - email
        - purpose
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [verify_email, reset_password]
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [driver, car_owner, admin]
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        is_email_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [created, role_selected, photo_uploaded, documents_uploaded, complete]
        profile_photo_url:
          type: string
          nullable: true
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Email is required